Removing deprecated code from the MySQL driver. Working on moving the remaining classes to the driver's new home.
